# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the REST documentation of Music Theory API",
    "version": "1.0.0",
    "title": "Music Theory API",
    "contact": {
      "email": "edipermadi@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://raw.githubusercontent.com/edipermadi/music-db/main/LICENSE"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/theory",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/chords": {
      "get": {
        "operationId": "ListChords",
        "tags": [
          "chord"
        ],
        "summary": "List chords",
        "description": "List simplified chords",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord quality identifier",
            "name": "chord_quality_id",
            "in": "query",
            "required": false,
            "minimum": 1,
            "type": "integer"
          },
          {
            "description": "Chord root pitch identifier",
            "name": "root_id",
            "in": "query",
            "required": false,
            "minimum": 1,
            "maximum": 12,
            "type": "integer"
          },
          {
            "description": "William Zeitler's chord number",
            "name": "zeitler_number",
            "in": "query",
            "required": false,
            "minimum": 1,
            "type": "integer"
          },
          {
            "description": "Ian Ring's chord number",
            "name": "ring_number",
            "in": "query",
            "required": false,
            "minimum": 1,
            "type": "integer"
          },
          {
            "description": "Chord cardinality, count of pitches in the chord",
            "name": "cardinality",
            "in": "query",
            "required": false,
            "minimum": 1,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ListChordsResponse"
            }
          }
        }
      }
    },
    "/chords/{chord_id}": {
      "get": {
        "operationId": "GetChord",
        "tags": [
          "chord"
        ],
        "summary": "Get chord",
        "description": "Get detailed chord information",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord identifier",
            "name": "chord_id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GetChordResponse"
            }
          }
        }
      }
    },
    "/chords/{chord_id}/quality": {
      "get": {
        "operationId": "GetChordQuality",
        "tags": [
          "chord"
        ],
        "summary": "Get chord quality",
        "description": "Get detailed chord quality",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord identifier",
            "name": "chord_id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/GetChordQualityResponse"
            }
          }
        }
      }
    },
    "/chords/{chord_id}/pitches": {
      "get": {
        "operationId": "ListChordPitches",
        "tags": [
          "chord"
        ],
        "summary": "List chord pitches",
        "description": "Get pitches within given chord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord identifier",
            "name": "chord_id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ListChordPitchesResponse"
            }
          }
        }
      }
    },
    "/chords/{chord_id}/scales": {
      "get": {
        "operationId": "ListChordScales",
        "tags": [
          "chord"
        ],
        "summary": "List chord scales",
        "description": "Get scales related to given chord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord identifier",
            "name": "chord_id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "type": "integer"
          },
          {
            "name": "tonic_id",
            "description": "Scale tonic pitch identifier",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          {
            "name": "zeitler_number",
            "description": "William Zeitler's scale number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "ring_number",
            "description": "Ian Ring's scale number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "perfection",
            "description": "Count of perfect notes",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 0
          },
          {
            "name": "imperfection",
            "description": "Count of imperfect notes",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 0
          },
          {
            "name": "balanced",
            "description": "Whether scale is balanced",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "rotational_symmetric",
            "description": "Whether scale is rotational symmetric",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "rotational_symmetry_level",
            "description": "Scale rotational symmetry level",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "palindromic",
            "description": "Whether scale is palindromic",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "cardinality",
            "description": "Scale cardinality, count of pitches in the scale",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "page",
            "description": "Page Number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "per_page",
            "description": "Page Size",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ListChordKeysResponse"
            }
          }
        }
      }
    },
    "/chords/{chord_id}/keys": {
      "get": {
        "operationId": "ListChordKeys",
        "tags": [
          "chord"
        ],
        "summary": "List chord keys",
        "description": "Get keys related to given chord",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Chord identifier",
            "name": "chord_id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "type": "integer"
          },
          {
            "name": "scale_id",
            "description": "Scale identifier identifier",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "tonic_id",
            "description": "Key tonic pitch identifier",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          {
            "name": "zeitler_number",
            "description": "William Zeitler's key number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "ring_number",
            "description": "Ian Ring's key number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "perfection",
            "description": "Count of perfect notes in the key",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 0
          },
          {
            "name": "imperfection",
            "description": "Count of imperfect notes in the key",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 0
          },
          {
            "name": "balanced",
            "description": "Whether the key is balanced",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "rotational_symmetric",
            "description": "Whether the key is rotational symmetric",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "rotational_symmetry_level",
            "description": "Key rotational symmetry level",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "palindromic",
            "description": "Whether key is palindromic",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "cardinality",
            "description": "Key cardinality, count of pitches in the key",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "page",
            "description": "Page Number",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1
          },
          {
            "name": "per_page",
            "description": "Page Size",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ListChordKeysResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ListChordsResponse": {
      "title": "List chords response",
      "type": "array",
      "items": {
        "$ref": "#/definitions/SimplifiedChord"
      },
      "minItems": 0,
      "maxItems": 12
    },
    "GetChordResponse": {
      "title": "Get chord response",
      "type": "object",
      "$ref": "#/definitions/DetailedChord"
    },
    "GetChordQualityResponse": {
      "title": "Get detailed chord quality response",
      "type": "object",
      "$ref": "#/definitions/DetailedChordQuality"
    },
    "ListChordPitchesResponse": {
      "title": "List of pitches within given chord",
      "type": "array",
      "items": {
        "$ref": "#/definitions/SimplifiedPitch"
      },
      "minItems": 0
    },
...
    "ListChordKeysResponse": {
      "title": "List of keys related to given chord",
      "type": "array",
      "items": {
        "$ref": "#/definitions/SimplifiedKey"
      },
      "minItems": 0
    },
...
    "SimplifiedPitch": {
      "title": "Simplified pitch information",
      "properties": {
        "id": {
          "$ref": "#/definitions/PitchId"
        },
        "name": {
          "$ref": "#/definitions/PitchName"
        }
      }
    },
...
    "PitchId": {
      "title": "Pitch identifier",
      "type": "integer",
      "minimum": 1,
      "maximum": 12,
      "example": 1,
      "readOnly": true
    },
    "PitchName": {
      "title": "Pitch name",
      "type": "string",
      "example": "CNatural",
      "readOnly": true
    },
...
    "SimplifiedChord": {
      "title": "Simplified chord information",
      "properties": {
        "id": {
          "$ref": "#/definitions/ChordId"
        },
        "name": {
          "$ref": "#/definitions/ChordName"
        }
      }
    },
    "DetailedChord": {
      "title": "Detailed chord information",
      "properties": {
        "id": {
          "$ref": "#/definitions/ChordId"
        },
        "name": {
          "$ref": "#/definitions/ChordName"
        },
        "zeitler_number": {
          "$ref": "#/definitions/ZeitlerChordNumber"
        },
        "ring_number": {
          "$ref": "#/definitions/RingChordNumber"
        },
        "quality": {
          "$ref": "#/definitions/SimplifiedChordQuality"
        },
        "root": {
          "$ref": "#/definitions/SimplifiedPitch"
        }
      }
    },
    "ChordId": {
      "title": "Chord identifier",
      "type": "integer",
      "minimum": 1,
      "example": 1,
      "readOnly": true
    },
    "ChordName": {
      "title": "Chord name",
      "type": "string",
      "example": "CNaturalMajor",
      "readOnly": true
    },
    "ZeitlerChordNumber": {
      "title": "Chord number in William Zeitler's numbering system",
      "description": "Value is determined by bitmap weight where CNatural is most significant bit and BNatural is the least significant bit. So CNatural number is 2048 while BNatural number is 1",
      "type": "integer",
      "minimum": 1,
      "readOnly": true
    },
    "RingChordNumber": {
      "title": "Chord number in Ian Ring's numbering system",
      "description": "Value is determined by bitmap weight where CNatural is least significant bit and BNatural is the most significant bit. So CNatural number is 1 while BNatural number is 2048",
      "type": "integer",
      "minimum": 1,
      "readOnly": true
    },
...
    "SimplifiedKey": {
      "title": "Simplified key",
      "properties": {
        "id": {
          "$ref": "#/definitions/KeyId"
        },
        "name": {
          "$ref": "#/definitions/KeyName"
        }
      }
    },
...
    "KeyId": {
      "title": "Key identifier",
      "type": "integer",
      "minimum": 1,
      "example": 1,
      "readOnly": true
    },
    "KeyName": {
      "title": "Key name",
      "type": "string",
      "example": "CNaturalMinoric",
      "readOnly": true
    },
...
    "SimplifiedChordQuality": {
      "title": "Simplified chord quality",
      "properties": {
        "id": {
          "$ref": "#/definitions/ChordQualityId"
        },
        "name": {
          "$ref": "#/definitions/ChordQualityName"
        }
      }
    },
    "DetailedChordQuality": {
      "title": "Simplified chord quality",
      "properties": {
        "id": {
          "$ref": "#/definitions/ChordQualityId"
        },
        "name": {
          "$ref": "#/definitions/ChordQualityName"
        },
        "zeitler_number": {
          "$ref": "#/definitions/ZeitlerChordQualityNumber"
        },
        "ring_number": {
          "$ref": "#/definitions/RingChordQualityNumber"
        },
        "cardinality": {
          "$ref": "#/definitions/ChordQualityCardinality"
        },
        "pitch_class": {
          "type": "array",
          "description": "List of pitch class",
          "items": {
            "description": "Pitch class number. 0 being CNatural, 11 being BNatural",
            "type": "number",
            "minimum": 0,
            "maximum": 11
          }
        },
        "interval_pattern": {
          "type": "array",
          "description": "Interval Pattern",
          "items": {
            "description": "Number of semitones between each notes",
            "type": "number",
            "minimum": 1
          }
        }
      }
    },
    "ChordQualityId": {
      "title": "Chord quality identifier",
      "type": "integer",
      "minimum": 1,
      "example": 1,
      "readOnly": true
    },
    "ChordQualityName": {
      "title": "Chord quality name",
      "type": "string",
      "example": "Major",
      "readOnly": true
    },
    "ZeitlerChordQualityNumber": {
      "title": "Chord quality number in William Zeitler's numbering system",
      "description": "Value is determined by bitmap weight where CNatural is most significant bit and BNatural is the least significant bit. So CNatural number is 2048 while BNatural number is 1",
      "type": "integer",
      "minimum": 1,
      "readOnly": true
    },
    "RingChordQualityNumber": {
      "title": "Chord quality number in Ian Ring's numbering system",
      "description": "Value is determined by bitmap weight where CNatural is least significant bit and BNatural is the most significant bit. So CNatural number is 1 while BNatural number is 2048",
      "type": "integer",
      "minimum": 1,
      "readOnly": true
    },
    "ChordQualityCardinality": {
      "title": "Chord quality cardinality",
      "description": "Count of note in a given chord quality",
      "type": "integer",
      "minimum": 1,
      "readOnly": true
    }
  }
}